pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devlution"
version = "0.1.0"
description = "AI-augmented developer workflow framework — from issue to PR with human supervision gates"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Thinkelution"},
]
keywords = ["ai", "developer-tools", "automation", "workflow", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "typer[all]>=0.9.0",
    "anthropic>=0.40.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pygithub>=2.0",
    "slack-bolt>=1.18",
    "langfuse>=2.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
devlution = "devlution.cli.main:app"

[project.urls]
Homepage = "https://github.com/thinkelution/devlution"
Repository = "https://github.com/thinkelution/devlution"
Issues = "https://github.com/thinkelution/devlution/issues"

[tool.setuptools.packages.find]
include = ["devlution*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
norecursedirs = ["tests/fixtures"]
